# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "TargetDesk API",
        "description": "API documentation for TargetDesk application",
        "contact": {
            "email": "admin@targetdesk.com"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost:8000/api",
            "description": "Local development server"
        }
    ],
    "paths": {
...
        "/v1/suppliers": {
            "get": {
                "tags": [
                    "Suppliers"
                ],
                "summary": "List all suppliers",
                "description": "Get a paginated list of all active suppliers",
                "operationId": "9ecdacd31b01411d71b4082fec58b598",
                "parameters": [
                    {
                        "name": "page",
                        "in": "query",
                        "description": "Page number",
                        "schema": {
                            "type": "integer",
                            "example": 1
                        }
                    },
                    {
                        "name": "per_page",
                        "in": "query",
                        "description": "Items per page (max 100)",
                        "schema": {
                            "type": "integer",
                            "example": 15
                        }
                    },
                    {
                        "name": "relation_type",
                        "in": "query",
                        "description": "Filter by relation type",
                        "schema": {
                            "type": "string",
                            "enum": [
                                "fournisseur",
                                "client_et_fournisseur"
                            ]
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Suppliers retrieved successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Fournisseurs récupérés avec succès"
                                        },
                                        "data": {
                                            "properties": {
                                                "suppliers": {
                                                    "type": "array",
                                                    "items": {
                                                        "type": "object"
                                                    }
                                                },
                                                "pagination": {
                                                    "type": "object"
                                                }
                                            },
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Suppliers"
                ],
                "summary": "Create a new supplier",
                "description": "Create a new supplier record with automatic ID generation",
                "operationId": "4cac2a9183d02b7967708ed1a636c0be",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "name",
                                    "type",
                                    "email"
                                ],
                                "properties": {
                                    "name": {
                                        "type": "string",
                                        "example": "Fournisseur ACME"
                                    },
                                    "type": {
                                        "type": "string",
                                        "enum": [
                                            "particulier",
                                            "entreprise"
                                        ],
                                        "example": "entreprise"
                                    },
                                    "email": {
                                        "type": "string",
                                        "format": "email",
                                        "example": "contact@acme-supplier.com"
                                    },
                                    "phone": {
                                        "type": "string",
                                        "example": "0123456789"
                                    },
                                    "address": {
                                        "type": "string",
                                        "example": "123 Rue du Fournisseur, 75001 Paris"
                                    },
                                    "siret": {
                                        "type": "string",
                                        "example": "12345678901234"
                                    },
                                    "sector": {
                                        "type": "string",
                                        "example": "Distribution"
                                    },
                                    "website": {
                                        "type": "string",
                                        "example": "https://acme-supplier.com"
                                    },
                                    "notes": {
                                        "type": "string",
                                        "example": "Notes importantes"
                                    },
                                    "relation_type": {
                                        "type": "string",
                                        "enum": [
                                            "fournisseur",
                                            "client_et_fournisseur"
                                        ],
                                        "example": "fournisseur"
                                    },
                                    "payment_terms": {
                                        "type": "string",
                                        "example": "30 jours net"
                                    },
                                    "delivery_delay": {
                                        "type": "integer",
                                        "example": 7
                                    },
                                    "currency": {
                                        "type": "string",
                                        "example": "EUR"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Supplier created successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Fournisseur créé avec succès"
                                        },
                                        "data": {
                                            "properties": {
                                                "id": {
                                                    "type": "integer",
                                                    "example": 1
                                                },
                                                "supplier_id": {
                                                    "type": "string",
                                                    "example": "FOUR-ABC123XY"
                                                },
                                                "name": {
                                                    "type": "string",
                                                    "example": "Fournisseur ACME"
                                                },
                                                "type": {
                                                    "type": "string",
                                                    "example": "entreprise"
                                                },
                                                "email": {
                                                    "type": "string",
                                                    "example": "contact@acme-supplier.com"
                                                }
                                            },
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validation error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": false
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Erreurs de validation"
                                        },
                                        "errors": {
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "sanctum": {
                "type": "http",
                "description": "Enter token in format: Bearer {token}",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
...
}